refactor: Move createJobs to @bundle-stats/utils
