+++ packages/bundle-stats/package.json
@@ -16,7 +16,7 @@
     "build": "babel src -d lib",
     "lint": "eslint .",
     "prepublishOnly": "npm run build",
-    "test": "jest src"
+    "test": "jest --passWithNoTests src "
   },
   "keywords": [
     "webpack",
