--- packages/bundle-stats/src/create-jobs.js
@@ -1,14 +1,0 @@
-import { createJob } from '@bundle-stats/utils';
-import { last, reverse } from 'lodash';
-
-export const createJobs = (sources) => {
-  const jobs = reverse([...sources]).reduce((agg, source, idx) => [
-    {
-      ...createJob({ webpack: { stats: source } }, last(agg)),
-      internalBuildNumber: (sources.length - idx),
-    },
-    ...agg,
-  ], []);
-
-  return jobs;
-};
